fix: Moved notifications.tsx file to apps/web/src/utils/notifications.tsx
