+++ apps/web/src/utils/notifications.tsx
@@ -1,7 +1,7 @@
 import { showNotification } from '@mantine/notifications';
-import { Check } from '../../design-system/icons/general/Check';
-import { ErrorIcon } from '../../design-system/icons';
-import { colors } from '../../design-system';
+import { Check } from '../design-system/icons/general/Check';
+import { ErrorIcon } from '../design-system/icons';
+import { colors } from '../design-system';
 
 export function successMessage(message: string) {
   showNotification({
